Move classes out of PowBlockDefinition
